+++ projects/ngx-collection/src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './collection';
 export * from './comparator';
 export * from './status.pipes';
+export * from './effect';
 export type * from './types';